refactor(system): remove redundant type cast

+++ system/utf8/substr.php
@@ -35,7 +35,7 @@ function _substr($str, $offset, $length = null)
     if ($offset > 0) {
         // PCRE repeating quantifiers must be less than 65536, so repeat when necessary
         $x = (int) ($offset / 65535);
-        $y = (int) ($offset % 65535);
+        $y = $offset % 65535;
         $regex .= ($x == 0) ? '' : ('(?:.{65535}){' . $x . '}');
         $regex .= ($y == 0) ? '' : ('.{' . $y . '}');
     }
@@ -50,7 +50,7 @@ function _substr($str, $offset, $length = null)
         $length = min($strlen - $offset, $length);
 
         $x = (int) ($length / 65535);
-        $y = (int) ($length % 65535);
+        $y = $length % 65535;
         $regex .= '(';
         $regex .= ($x == 0) ? '' : ('(?:.{65535}){' . $x . '}');
         $regex .= '.{' . $y . '})';
@@ -58,7 +58,7 @@ function _substr($str, $offset, $length = null)
     // Find length from the right (negative length)
     else {
         $x = (int) (-$length / 65535);
-        $y = (int) (-$length % 65535);
+        $y = -$length % 65535;
         $regex .= '(.*)';
         $regex .= ($x == 0) ? '' : ('(?:.{65535}){' . $x . '}');
         $regex .= '.{' . $y . '}';
